backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0

# ASGI Server
uvicorn[standard]==0.27.0

# Database ORM and Driver
sqlmodel==0.0.14
asyncpg==0.29.0

# Authentication and Security
PyJWT==2.8.0
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# CORS Middleware (included in FastAPI but explicit for clarity)
# fastapi already includes starlette which has CORS support
